MET-2046: Add server_time to order update calls
- get rid of the front-end order call
- unify the order observation
- send server_time
