--- Observer/UpdateOrder.php
@@ -1,33 +1,0 @@
-<?php
-
-namespace Metrilo\Analytics\Observer;
-
-use Magento\Framework\Event\Observer;
-use Magento\Framework\Event\ObserverInterface;
-
-class UpdateOrder implements ObserverInterface
-{
-
-    private $_helper;
-
-    /**
-     * @param \Metrilo\Analytics\Helper\Data $helper
-     */
-    public function __construct(
-        \Metrilo\Analytics\Helper\Data $helper
-    ) {
-        $this->_helper = $helper;
-    }
-
-    /**
-     * Trigger on save Order
-     *
-     * @param  \Magento\Framework\Event\Observer $observer
-     * @return void
-     */
-    public function execute(Observer $observer)
-    {
-        $order = $observer->getEvent()->getOrder();
-        $this->_helper->callBatchApi($order->getStoreId(), [$order]);
-    }
-}
