+++ etc/events.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 /**
- * Register adminhtml events
+ * Register global events
  *
  * @author Miroslav Petrov <miro91tn@gmail.com>
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Event/etc/events.xsd">
-    <event name="admin_system_config_changed_section_metrilo_analytics">
-        <observer name="metrilo_config_change" instance="Metrilo\Analytics\Observer\Config"/>
+    <event name="sales_order_save_after">
+        <observer name="metrilo_event_order" instance="Metrilo\Analytics\Observer\Order"/>
     </event>
 </config>
