+++ etc/frontend/events.xml
@@ -7,12 +7,6 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Event/etc/events.xsd">
-    <event name="checkout_onepage_controller_success_action">
-        <observer name="metrilo_event_order" instance="Metrilo\Analytics\Observer\Order"/>
-    </event>
-    <event name="multishipping_checkout_controller_success_action">
-        <observer name="metrilo_event_order" instance="Metrilo\Analytics\Observer\Order"/>
-    </event>
     <event name="customer_login">
         <observer name="metrilo_event_login" instance="Metrilo\Analytics\Observer\CustomerLogin" />
     </event>
